Put error message related stuff into their own file
