+++ client-v2/src/app/components/atoms/input/input.component.ts
@@ -1,26 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core'
 import { AbstractControl, FormControl, FormGroup } from '@angular/forms'
 import { stripNonWordChars } from 'src/app/utils'
-
-export interface FormValidationErrorMessages {
-    required: string
-    minLength: string
-    maxLength: string
-    min: string
-    max: string
-    email: string
-    pattern: string
-}
-
-const defaultErrorMessages: FormValidationErrorMessages = {
-    required: '$name is required.',
-    email: 'Must be a valid email address.',
-    minLength: 'Must be at least $value characters long.',
-    maxLength: 'Must not be longer than $value characters.',
-    min: 'Must be at least $value.',
-    max: 'Must not be higher than $value.',
-    pattern: 'Must match the pattern.',
-}
+import { defaultErrorMessages } from './error-messages'
+import { FormValidationErrorMessages } from './error-messages'
 
 const errorValueKeyMap = {
     minLength: 'requiredLength',
